Serve static ndjson file
