+++ bcda/main.go
@@ -1,24 +1,19 @@
 package main
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log"
 	"net/http"
 	"os"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/CMSgov/bcda-app/bcda/auth"
-	"github.com/CMSgov/bcda-app/bcda/database"
-	"github.com/CMSgov/bcda-app/models"
 	"github.com/urfave/cli"
 
 	"github.com/bgentry/que-go"
 	"github.com/dgrijalva/jwt-go"
-	"github.com/go-chi/chi"
 	"github.com/jackc/pgx"
 	"github.com/pborman/uuid"
 )
@@ -51,138 +46,6 @@ func claimsFromToken(token *jwt.Token) (jwt.MapClaims, error) {
 		return claims, nil
 	}
 	return jwt.MapClaims{}, errors.New("Error determining token claims")
-}
-
-func bulkRequest(w http.ResponseWriter, r *http.Request) {
-	var (
-		claims jwt.MapClaims
-		err    error
-	)
-
-	db := database.GetDbConnection()
-	defer db.Close()
-
-	t := r.Context().Value("token")
-	if token, ok := t.(*jwt.Token); ok && token.Valid {
-		claims, err = claimsFromToken(token)
-		if err != nil {
-			log.Fatal(err)
-		}
-	}
-
-	acoId, _ := claims["aco"].(string)
-	userId, _ := claims["sub"].(string)
-
-	scheme := "http"
-	if r.TLS != nil {
-		scheme = "https"
-	}
-
-	newJob := models.Job{
-		AcoID:      uuid.Parse(acoId),
-		UserID:     uuid.Parse(userId),
-		RequestURL: fmt.Sprintf("%s://%s%s", scheme, r.Host, r.URL),
-		Status:     "Pending",
-	}
-	if err := newJob.Insert(db); err != nil {
-		log.Fatal(err)
-	}
-
-	args, err := json.Marshal(jobEnqueueArgs{
-		ID:     newJob.ID,
-		AcoID:  acoId,
-		UserID: userId,
-	})
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	j := &que.Job{
-		Type: "ProcessJob",
-		Args: args,
-	}
-	if err = qc.Enqueue(j); err != nil {
-		log.Fatal(err)
-	}
-
-	w.Header().Set("Content-Location", fmt.Sprintf("%s://%s/api/v1/jobs/%d", scheme, r.Host, newJob.ID))
-	w.WriteHeader(http.StatusAccepted)
-}
-
-func jobStatus(w http.ResponseWriter, r *http.Request) {
-	jobID := chi.URLParam(r, "jobId")
-	db := database.GetDbConnection()
-	defer db.Close()
-
-	i, err := strconv.Atoi(jobID)
-	if err != nil {
-		log.Print(err)
-		http.Error(w, http.StatusText(400), 400)
-		return
-	}
-
-	job, err := models.JobByID(db, i)
-	if err != nil {
-		log.Print(err)
-		http.Error(w, http.StatusText(404), 404)
-		return
-	}
-
-	switch job.Status {
-	case "Pending":
-		fallthrough
-	case "In Progress":
-		w.Header().Set("X-Progress", job.Status)
-		w.WriteHeader(http.StatusAccepted)
-	case "Failed":
-		http.Error(w, http.StatusText(500), 500)
-	case "Completed":
-		w.Header().Set("Content-Type", "application/json")
-
-		fi := fileItem{
-			Type: "Patient",
-			URL:  "http://serverpath2/patient_file_1.ndjson",
-		}
-
-		rb := bulkResponseBody{
-			TransactionTime:     time.Now(),
-			RequestURL:          job.RequestURL,
-			RequiresAccessToken: true,
-			Files:               []fileItem{fi},
-			Errors:              []fileItem{},
-		}
-
-		jsonData, err := json.Marshal(rb)
-		if err != nil {
-			http.Error(w, http.StatusText(500), 500)
-			return
-		}
-
-		_, err = w.Write([]byte(jsonData))
-		if err != nil {
-			http.Error(w, http.StatusText(500), 500)
-			return
-		} else {
-			w.WriteHeader(http.StatusOK)
-		}
-	}
-}
-
-func getToken(w http.ResponseWriter, r *http.Request) {
-	authBackend := auth.InitAuthBackend()
-
-	// Generates a token for fake user and ACO combination
-	token, err := authBackend.GenerateToken(
-		"82503A18-BF3B-436D-BA7B-BAE09B7FFD2F",
-		"DBBD1CE1-AE24-435C-807D-ED45953077D3",
-	)
-	if err != nil {
-		log.Fatal(err)
-	}
-	_, err = w.Write([]byte(token))
-	if err != nil {
-		log.Fatal(err)
-	}
 }
 
 func main() {
