+++ bcda/router.go
@@ -29,5 +29,9 @@ func NewRouter() http.Handler {
 			r.Get("/token", getToken)
 		}
 	})
+	r.Route("/data", func(r chi.Router) {
+		// With auth to check ACO ID of requested file against token
+		r.Get("/DBBD1CE1-AE24-435C-807D-ED45953077D3.ndjson", serveData)
+	})
 	return r
 }
